requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
idna==3.11
py-cpuinfo==9.0.0
random-username==1.0.2
requests==2.32.5
urllib3==2.5.0